Introducing a bad pattern for getting submodule CLI commands included into the main command chain.  Updated tests to comply with the refactoring.
gh-#3
